hw07/src/main/java/ru/otus/spring/hw07/services/impl/GenresServiceImpl.java
package ru.otus.spring.hw07.services.impl;

import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import ru.otus.spring.hw07.domain.Genre;
import ru.otus.spring.hw07.repository.GenreRepository;
import ru.otus.spring.hw07.services.GenresService;

import java.util.List;
import java.util.Optional;

@Service
@RequiredArgsConstructor
public class GenresServiceImpl implements GenresService {

    private final GenreRepository genreRepository;

    @Override
    public List<Genre> findAll() {
        return genreRepository.findAll();
    }

    @Override
    public List<Genre> findAllById(Iterable<Long> ids) {
        return genreRepository.findAllById(ids);
    }

    @Override
    public Optional<Genre> findById(long id) {
        return genreRepository.findById(id);
    }

    @Override
    public Genre save(Genre genre) {
        return genreRepository.save(genre);
    }

    @Override
    public void deleteById(long id) {
        genreRepository.deleteById(id);
    }
}
